+++ fulltext/backends/__epub.py
@@ -6,7 +6,7 @@ from ebooklib import epub
 from bs4 import BeautifulSoup
 from six import StringIO
 
-from fulltext import BaseBackend
+from fulltext.util import BaseBackend
 from fulltext.util import assert_cmd_exists
 from fulltext.util import exiftool_title
