+++ fulltext/backends/__html.py
@@ -7,7 +7,7 @@ from bs4 import BeautifulSoup
 from six import StringIO
 from six import PY3
 
-from fulltext import BaseBackend
+from fulltext.util import BaseBackend
 
 
 class Backend(BaseBackend):
