+++ fulltext/backends/__pptx.py
@@ -4,7 +4,7 @@ import os
 import pptx
 from six import StringIO
 
-from fulltext import BaseBackend
+from fulltext.util import BaseBackend
 from fulltext.util import assert_cmd_exists
 from fulltext.util import exiftool_title
