+++ fulltext/backends/__ps.py
@@ -4,7 +4,7 @@
 from __future__ import absolute_import
 
 from fulltext.util import run, assert_cmd_exists, exiftool_title
-from fulltext import BaseBackend
+from fulltext.util import BaseBackend
 
 
 class Backend(BaseBackend):
