+++ fulltext/backends/__rtf.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 
 from fulltext.util import run, assert_cmd_exists, exiftool_title
-from fulltext import BaseBackend
+from fulltext.util import BaseBackend
 
 
 class Backend(BaseBackend):
