+++ fulltext/backends/__xml.py
@@ -7,7 +7,7 @@ import bs4
 from six import StringIO
 from six import PY3
 
-from fulltext import BaseBackend
+from fulltext.util import BaseBackend
 
 
 class Backend(BaseBackend):
