+++ fulltext/backends/__zip.py
@@ -6,7 +6,7 @@ import zipfile
 from six import StringIO
 from contextlib2 import ExitStack
 
-from fulltext import BaseBackend, get
+from fulltext.util import BaseBackend, get
 
 
 LOGGER = logging.getLogger(__name__)
